fix: mount options modal in sandbox client
