+++ sandbox/src/Controls.tsx
@@ -1,7 +1,6 @@
 import {createPortal} from "react-dom";
 import {Button, Modal} from "react-bootstrap";
 import {useState} from "react";
-import OptionsApp from "@options/src/App.tsx";
 import TriggerTester from "./TriggerTester.tsx";
 import packageAssistant from "./scenario/package-assistant.ts";
 import killCounterDemo from "./scenario/kill-counter-demo.ts";
@@ -46,7 +45,6 @@ function runDemo(demoName: string) {
 
 export function Controls() {
     const [showTester, setShowTester] = useState(false);
-    const [showOptions, setShowOptions] = useState(false);
     return (
         <>
             {createPortal(
@@ -153,13 +151,6 @@ export function Controls() {
                     >
                         Invite Demo
                     </Button>
-                    <Button
-                        size="sm"
-                        variant="secondary"
-                        onClick={() => setShowOptions(true)}
-                    >
-                        Options
-                    </Button>
                 </div>,
                 document.getElementById('sandbox-buttons')!,
             )}
@@ -169,14 +160,6 @@ export function Controls() {
                 </Modal.Header>
                 <Modal.Body>
                     <TriggerTester/>
-                </Modal.Body>
-            </Modal>
-            <Modal show={showOptions} onHide={() => setShowOptions(false)} fullscreen>
-                <Modal.Header closeButton>
-                    <Modal.Title>Options</Modal.Title>
-                </Modal.Header>
-                <Modal.Body>
-                    <OptionsApp />
                 </Modal.Body>
             </Modal>
         </>
